Abstracted PendingTx to its own class for better clarity
